Catch verify options error(s)
I think it should catch the errors thrown while verifying options instead of let them block the process, so if the script is used inside other scripts it will fail without stopping global execution, for example if the host parameter is missing or some options where set incorrectly.

Also I think for callback-based (ie. asynchronous) code, if the first argument of the callback is err, like in this case it should always return errors instead of throwing them directly.